ld/startup: unify startup for AXI and FLASH builds, code/data to ITCM/DTCM
- Simplified startup: single Reset_Handler now supports both RAM and flash boot.
- Removed need for separate dtcm-only startup variant.
  Both AXI and FLASH builds now share the same startup.c
Changes enable code and data placement in specific memory regions:
- Functions marked `__attribute__((section(".itcm")))` now run from ITCM (fast, zero-wait)
- Variables in `.dtcm_data` or `.dtcm_bss` live in DTCM (deterministic access)

+++ src/h750/dwt/dwt.c
@@ -30,10 +30,11 @@ double dwt_cycles_to_s(uint32_t cycles)
 
 double dwt_cycles_to_ms(uint32_t cycles)
 {
-    return (double)cycles/((double)sysclk_hz/1000.0);
+    return (double)cycles/((double)sysclk_hz/(double)1000.0);
 }
 
 double dwt_cycles_to_us(uint32_t cycles)
 {
-    return (double)cycles/((double)sysclk_hz/1000000.0);
+    return (double)cycles/((double)sysclk_hz/(double)1000000.0);
 }
+
